Move job_applications route and view from jobs blueprint to applications blueprint; move template for job_applications to applications directory, update file paths, route and function names accordingly to better reflect the purpose of job_applications
