+++ app/routes/jobs.py
@@ -96,19 +96,3 @@ def delete_job(job_id):
     db.session.commit()
 
     return redirect(url_for('job_blueprint.jobs'))
-
-@job_blueprint.route('/<job_id>/applications', methods=['GET'])
-@login_required
-@admin_permission.require()
-def job_applications(job_id):
-    job = Job.query.filter_by(id=job_id).first()
-    applications = Application.query.filter_by(job_id=job_id).all()
-
-    for application in applications:
-        user = User.query.filter_by(id=application.user_id).first()
-        if user:
-            print(f"Application ID: {application.id}, Created by: {user.name}")
-        else:
-            print(f"Application ID: {application.id}, User not found for user_id: {application.user_id}")
-
-    return render_template('jobs/job_applications.html', job=job, applications=applications, primary_title='Job Applications')
