+++ app/templates/applications/job_applications.html
